#1701: extract routing/protocol types into types_routing.go
Moves the routing-options, policy-options, dynamic-protocol, tunnel, and
VRF type definitions verbatim out of types.go into types_routing.go
(same package config):

  PolicyOptionsConfig, ASPathDef, CommunityDef, PrefixList,
  PolicyStatement, PolicyTerm, RouteFilter, RoutingOptionsConfig,
  GenerateRoute, RibGroup, NextHopEntry, StaticRoute, ProtocolsConfig,
  LLDPConfig(+Interface), OSPFv3*, RIPConfig, ISISConfig(+Interface),
  RAInterfaceConfig, RAPrefix, OSPFConfig/Area/VirtualLink/Interface,
  BGPConfig, BGPNeighbor, TunnelConfig, RoutingInstanceConfig.

IPsec/IKE types are NOT here despite the routing adjacency: they nest
under SecurityConfig.IPsec and live at [edit security ike/ipsec], so they
go in types_security.go (per the plan's AGY/Codex re-bucket). Pure
byte-identical motion; consumers untouched. Verified: go build ./...
clean, vet clean, tests pass.
