+++ pkg/config/types.go
@@ -281,64 +281,6 @@ func IRBToBridge(bds []*BridgeDomainConfig) map[string]string {
 	return m
 }
 
-// PolicyOptionsConfig holds prefix-lists, communities, as-paths, and policy-statements for routing control.
-type PolicyOptionsConfig struct {
-	PrefixLists      map[string]*PrefixList
-	Communities      map[string]*CommunityDef
-	ASPaths          map[string]*ASPathDef
-	PolicyStatements map[string]*PolicyStatement
-}
-
-// ASPathDef defines a named AS-path regular expression for route matching.
-type ASPathDef struct {
-	Name  string
-	Regex string // AS-path regex pattern (e.g. "65000", "65[0-9]+")
-}
-
-// CommunityDef defines a named BGP community with member values.
-type CommunityDef struct {
-	Name    string
-	Members []string // e.g. "65000:100", "no-export", "no-advertise"
-}
-
-// PrefixList defines a named list of IP prefixes.
-type PrefixList struct {
-	Name     string
-	Prefixes []string // CIDR entries ("10.0.0.0/8", "2001:db8::/32")
-}
-
-// PolicyStatement defines a routing policy with terms.
-type PolicyStatement struct {
-	Name          string
-	Terms         []*PolicyTerm
-	DefaultAction string // "accept", "reject", or "" (implicit reject)
-}
-
-// PolicyTerm is a single match+action clause within a policy-statement.
-type PolicyTerm struct {
-	Name            string
-	FromProtocol    string         // "direct", "static", "bgp", "ospf"
-	PrefixList      string         // from prefix-list <name>
-	FromCommunity   string         // from community <name> (match against community-list)
-	FromASPath      string         // from as-path <name> (match against as-path access-list)
-	RouteFilters    []*RouteFilter // prefix matching
-	Action          string         // "accept", "reject"
-	NextHop         string         // then next-hop (e.g. "peer-address", "self", IP)
-	LoadBalance     string         // then load-balance (e.g. "consistent-hash", "per-packet")
-	LocalPreference int            // BGP local-preference (0 = not set)
-	Metric          int            // BGP MED/metric (0 = not set)
-	MetricType      int            // OSPF metric type (1 or 2, 0 = not set)
-	Community       string         // BGP community to set (e.g. "65000:100")
-	Origin          string         // BGP origin: "igp", "egp", "incomplete"
-}
-
-// RouteFilter matches a prefix with a match type.
-type RouteFilter struct {
-	Prefix    string // CIDR ("192.168.50.0/24")
-	MatchType string // "exact", "longer", "orlonger", "upto"
-	UptoLen   int    // for "upto" match type
-}
-
 // SchedulerConfig defines a time-based policy scheduler.
 type SchedulerConfig struct {
 	Name      string
@@ -1352,246 +1294,6 @@ type Application struct {
 	Description       string
 }
 
-// RoutingOptionsConfig holds static routing configuration.
-type RoutingOptionsConfig struct {
-	StaticRoutes              []*StaticRoute
-	Inet6StaticRoutes         []*StaticRoute // rib inet6.0 static routes
-	GenerateRoutes            []*GenerateRoute
-	ForwardingTableExport     string // forwarding-table { export <policy>; }
-	AutonomousSystem          uint32 // autonomous-system <number>
-	RibGroups                 map[string]*RibGroup
-	InterfaceRoutesRibGroup   string // global interface-routes { rib-group inet <name>; }
-	InterfaceRoutesRibGroupV6 string // global interface-routes { rib-group inet6 <name>; }
-}
-
-// GenerateRoute defines a Junos generate (aggregate) route.
-// In FRR, these become blackhole/reject static routes or BGP aggregate-address.
-type GenerateRoute struct {
-	Prefix  string // route prefix (e.g. "192.168.0.0/16")
-	Policy  string // contributing route policy (optional)
-	Discard bool   // discard traffic to this route (blackhole)
-}
-
-// RibGroup defines a RIB group for route sharing between routing instances.
-type RibGroup struct {
-	Name       string
-	ImportRibs []string // import-rib [ rib1 rib2 ... ]
-}
-
-// NextHopEntry defines a single next-hop for a static route.
-type NextHopEntry struct {
-	Address   string // IP address (e.g. "10.0.1.1" or "fe80::1")
-	Interface string // outgoing interface (for IPv6 link-local)
-}
-
-// StaticRoute defines a single static route.
-type StaticRoute struct {
-	Destination string         // CIDR: "10.0.0.0/8" or "::/0"
-	NextHops    []NextHopEntry // multiple next-hops = ECMP
-	Discard     bool           // null route (blackhole)
-	Preference  int            // route preference (admin distance), default 5
-	NextTable   string         // routing instance name for inter-VRF route leaking (e.g. "Comcast.inet.0" → "Comcast")
-}
-
-// ProtocolsConfig holds dynamic routing protocol configuration.
-type ProtocolsConfig struct {
-	OSPF                *OSPFConfig
-	OSPFv3              *OSPFv3Config
-	BGP                 *BGPConfig
-	RIP                 *RIPConfig
-	ISIS                *ISISConfig
-	RouterAdvertisement []*RAInterfaceConfig
-	LLDP                *LLDPConfig
-}
-
-// LLDPConfig holds LLDP (Link Layer Discovery Protocol) configuration.
-type LLDPConfig struct {
-	Interfaces     []LLDPInterface // interfaces to enable LLDP on
-	Interval       int             // transmit interval in seconds (0 = default 30)
-	HoldMultiplier int             // hold multiplier (0 = default 4)
-	Disable        bool            // globally disable LLDP
-}
-
-// LLDPInterface holds per-interface LLDP configuration.
-type LLDPInterface struct {
-	Name    string
-	Disable bool // per-interface disable
-}
-
-// OSPFv3Config holds OSPFv3 (IPv6 OSPF) routing configuration.
-type OSPFv3Config struct {
-	RouterID string
-	Areas    []*OSPFv3Area
-	Export   []string
-}
-
-// OSPFv3Area defines an OSPFv3 area.
-type OSPFv3Area struct {
-	ID         string // "0.0.0.0" (backbone) or area number
-	Interfaces []*OSPFv3Interface
-}
-
-// OSPFv3Interface defines an interface participating in OSPFv3.
-type OSPFv3Interface struct {
-	Name    string
-	Passive bool
-	Cost    int
-}
-
-// RIPConfig holds RIP routing configuration.
-type RIPConfig struct {
-	Interfaces   []string // interfaces participating in RIP
-	Passive      []string // passive interfaces (receive only)
-	Redistribute []string // "connected", "static", "ospf"
-	AuthKey      string   // authentication key/password
-	AuthType     string   // "md5" or "simple"
-}
-
-// ISISConfig holds IS-IS routing configuration.
-type ISISConfig struct {
-	NET             string // ISO NET address (e.g. "49.0001.0100.0000.0001.00")
-	Level           string // "level-1", "level-2", "level-1-2" (default "level-2")
-	Interfaces      []*ISISInterface
-	Export          []string // "connected", "static", etc.
-	AuthKey         string   // area-level authentication key
-	AuthType        string   // "md5" or "simple" (plaintext)
-	WideMetricsOnly bool     // use wide (32-bit) metrics
-	Overload        bool     // set overload bit
-}
-
-// ISISInterface defines an interface participating in IS-IS.
-type ISISInterface struct {
-	Name          string
-	Level         string // override per-interface
-	Passive       bool
-	Metric        int    // 0 = default
-	AuthKey       string // per-interface authentication key
-	AuthType      string // "md5" or "simple"
-	BFD           bool   // enable BFD on this interface
-	BFDInterval   int    // BFD minimum-interval in ms (0 = default)
-	BFDMultiplier int    // BFD detect-multiplier (0 = default)
-}
-
-// RAInterfaceConfig configures Router Advertisement on an interface.
-type RAInterfaceConfig struct {
-	Interface       string
-	ManagedConfig   bool   // managed-configuration (M flag)
-	OtherStateful   bool   // other-stateful-configuration (O flag)
-	Preference      string // "high", "medium", "low" (default: medium)
-	DefaultLifetime int    // seconds, 0 = default (1800)
-	MaxAdvInterval  int    // seconds, 0 = default (600)
-	MinAdvInterval  int    // seconds, 0 = default (200)
-	Prefixes        []*RAPrefix
-	DNSServers      []string // recursive DNS server addresses
-	NAT64Prefix     string   // PREF64 prefix (e.g. "64:ff9b::/96")
-	NAT64PrefixLife int      // PREF64 lifetime in seconds (0 = default)
-	LinkMTU         int      // advertised link MTU, 0 = omit
-	SourceLinkLocal string   // explicit link-local to use as RA source (overrides auto-selected)
-}
-
-// RAPrefix defines a prefix advertised via RA.
-type RAPrefix struct {
-	Prefix        string // CIDR notation
-	OnLink        bool   // on-link flag (default true)
-	Autonomous    bool   // SLAAC autonomous flag (default true)
-	ValidLifetime int    // seconds, 0 = default (2592000 = 30 days)
-	PreferredLife int    // seconds, 0 = default (604800 = 7 days)
-}
-
-// OSPFConfig holds OSPF routing configuration.
-type OSPFConfig struct {
-	RouterID           string // e.g. "10.0.0.1"
-	ReferenceBandwidth int    // Mbps for auto-cost calculation (0 = FRR default 100)
-	PassiveDefault     bool   // all interfaces passive by default
-	Areas              []*OSPFArea
-	Export             []string // export policy names (future)
-}
-
-// OSPFArea defines an OSPF area.
-type OSPFArea struct {
-	ID           string // "0.0.0.0" (backbone) or area number
-	AreaType     string // "stub", "nssa", "" (normal)
-	NoSummary    bool   // stub/nssa no-summary (totally stubby)
-	Interfaces   []*OSPFInterface
-	VirtualLinks []*OSPFVirtualLink
-}
-
-// OSPFVirtualLink defines a virtual link to the backbone through a transit area.
-type OSPFVirtualLink struct {
-	NeighborID  string // router-id of ABR at the other end
-	TransitArea string // the area this virtual-link traverses (the area ID where it's configured)
-}
-
-// OSPFInterface defines an interface participating in OSPF.
-type OSPFInterface struct {
-	Name          string
-	Passive       bool   // passive interface (no hello)
-	NoPassive     bool   // override passive-default (explicitly active)
-	Cost          int    // OSPF cost, 0 = default
-	NetworkType   string // "point-to-point", "broadcast", "" (default)
-	AuthType      string // "md5", "simple", "" (none)
-	AuthKey       string // authentication key/password
-	AuthKeyID     int    // key-id for MD5 (1-255)
-	BFD           bool   // enable BFD on this interface
-	BFDInterval   int    // BFD minimum-interval in ms (0 = default)
-	BFDMultiplier int    // BFD detect-multiplier (0 = default)
-}
-
-// BGPConfig holds BGP routing configuration.
-type BGPConfig struct {
-	LocalAS              uint32
-	RouterID             string
-	ClusterID            string // route reflector cluster ID
-	GracefulRestart      bool   // enable graceful restart
-	Multipath            int    // maximum equal-cost paths (0 = disabled)
-	MultipathMultipleAS  bool   // allow multipath across different ASes
-	LogNeighborChanges   bool   // log neighbor state transitions
-	Dampening            bool   // enable route flap dampening
-	DampeningHalfLife    int    // half-life in minutes (default 15)
-	DampeningReuse       int    // reuse threshold (default 750)
-	DampeningSuppress    int    // suppress threshold (default 2000)
-	DampeningMaxSuppress int    // max suppress time in minutes (default 60)
-	Neighbors            []*BGPNeighbor
-	Export               []string // "connected", "static", "ospf", etc.
-}
-
-// BGPNeighbor defines a BGP peer.
-type BGPNeighbor struct {
-	Address              string // peer IP
-	PeerAS               uint32
-	Description          string
-	MultihopTTL          int      // 0 = directly connected
-	Export               []string // per-group export policies (route-map out)
-	FamilyInet           bool     // activate under address-family ipv4 unicast
-	FamilyInet6          bool     // activate under address-family ipv6 unicast
-	GroupName            string   // BGP group name (for display)
-	AuthPassword         string   // TCP MD5 password for BGP session
-	BFD                  bool     // enable BFD for this neighbor
-	BFDInterval          int      // BFD minimum interval in ms (0 = default 300)
-	BFDMultiplier        int      // BFD detect-multiplier (0 = default 3)
-	RouteReflectorClient bool     // mark as route-reflector client
-	DefaultOriginate     bool     // advertise default route to this neighbor
-	AllowASIn            int      // allow own AS in path N times (0 = disabled)
-	RemovePrivateAS      bool     // strip private AS numbers from updates
-	PrefixLimitInet      int      // max IPv4 prefixes (0 = unlimited)
-	PrefixLimitInet6     int      // max IPv6 prefixes (0 = unlimited)
-}
-
-// TunnelConfig defines a GRE, IPIP, or other tunnel interface.
-type TunnelConfig struct {
-	Name            string   // Linux interface name (e.g. "gr-0-0-0", "ip-0-0-0")
-	Mode            string   // "gre" or "ipip"
-	Source          string   // local tunnel endpoint IP
-	Destination     string   // remote tunnel endpoint IP
-	Key             uint32   // GRE key, 0 = none
-	TTL             int      // tunnel TTL, 0 = default 64
-	Addresses       []string // IPs to assign to tunnel interface (CIDR)
-	RoutingInstance string   // destination routing-instance (VRF)
-	Keepalive       int      // keepalive interval in seconds (0 = disabled)
-	KeepaliveRetry  int      // number of missed keepalives before declaring down (0 = default 3)
-	AnchorOnly      bool     // create a dummy anchor instead of a kernel tunnel device
-}
-
 // TunnelNameMap returns a mapping from Junos interface reference (e.g. "gr-0/0/0.0")
 // to the Linux tunnel interface name. For tunnel interfaces with per-unit tunnel config,
 // unit 0 uses the base Linux name, unit N>0 appends "uN".
@@ -1706,22 +1408,4 @@ type IPsecVPN struct {
 	DFBit            string // "copy", "set", "clear"
 	EstablishTunnels string // "immediately", "on-traffic"
 	TrafficSelectors map[string]*IPsecTrafficSelector
-}
-
-// RoutingInstanceConfig represents a VRF-based routing instance.
-type RoutingInstanceConfig struct {
-	Name                      string
-	Description               string
-	InstanceType              string         // "virtual-router" or "vrf"
-	Interfaces                []string       // interfaces belonging to this instance
-	StaticRoutes              []*StaticRoute // per-instance static routes
-	Inet6StaticRoutes         []*StaticRoute // per-instance rib inet6.0 static routes
-	OSPF                      *OSPFConfig    // per-instance OSPF (optional)
-	OSPFv3                    *OSPFv3Config  // per-instance OSPFv3 (optional)
-	BGP                       *BGPConfig     // per-instance BGP (optional)
-	RIP                       *RIPConfig     // per-instance RIP (optional)
-	ISIS                      *ISISConfig    // per-instance IS-IS (optional)
-	TableID                   int            // Linux kernel routing table number (auto-assigned)
-	InterfaceRoutesRibGroup   string         // interface-routes { rib-group inet <name>; }
-	InterfaceRoutesRibGroupV6 string         // interface-routes { rib-group inet6 <name>; }
 }
